Renamed the model validator file to formValidators and added an extension type for IFormController
