+++ source/components/input/input.ts
@@ -9,7 +9,7 @@ import { services } from 'typescript-angular-utilities';
 import __validation = services.validation;
 import __object = services.object;
 
-import { INgModelValidator } from '../../types/ngModelValidator';
+import { INgModelValidator } from '../../types/formValidators';
 import { directiveName as requiredDirectiveName, RequiredController } from '../../behaviors/required/required';
 import {
 	IComponentValidator,
