+++ source/types/types.module.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import * as ngModel from './ngModelValidator';
+import * as ngModel from './formValidators';
 import * as viewData from './viewData';
 
 export { ngModel, viewData };
